Split foundry-mcp compendium-cache.ts (800 LOC) into 4 modules
The single file mixed three concerns: cache lifecycle (warm + storage),
read-side filter logic, and per-field readers for pf2e document
shapes. Splitting them clarifies each module's job and lets filter +
facet aggregation be unit-tested without instantiating a cache.

New structure:

- `compendium-types.ts` — shared interfaces (CompendiumDocument,
  EnrichedMatch, ItemPrice, SearchOptions, CachedPack, CompendiumCacheStats,
  SendCommand). Sits at the bottom of the local module graph; cycle-free.
- `compendium-extractors.ts` — every `extract*` field reader plus
  `bucketUsage`, `score`, and the pf2e trait constants
  (CREATURE_TYPE_TRAITS, TRADITION_TRAITS, USAGE_PREFIX_BUCKETS).
  Defensively-typed pure functions over `unknown` slices of pf2e docs.
- `compendium-search.ts` — pure `runFilter(packs, opts)` and
  `aggregateFacets(packs, opts)`. Both walk a `CachedPack[]` and call
  the extractors; testable without a cache instance.
- `compendium-cache.ts` — `CompendiumCache` class only. Owns warm
  orchestration, in-memory storage, hit/miss/byte stats, and the small
  public API (`search`, `facets`, `getDocument`, etc.). Delegates to
  the search module after looking up packs.

Drive-by simplification: the duplicated "resolve packIds → CachedPack[]
or null" prelude that `search()` and `facets()` shared is now a private
`lookup()` method. Same miss-bookkeeping semantics, less repetition.

Public surface unchanged. The cache module re-exports
`CompendiumDocument`, `SendCommand`, etc., so existing imports
(`import { CompendiumCache, type CompendiumDocument } from
'./compendium-cache.js'`) keep working — `test/compendium-cache.test.ts`
needs no changes and passes through.

Bucket C-a of the cleanup audit.
